+++ components/PageSections/contactPage/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+import AskQuesList from "./askedQues/askedQuesData";
+import ContactForm from "./contactForm";
+
+const ContactPage = () => {
+  return (
+    <>
+      <ContactForm />
+      <AskQuesList />
+    </>
+  );
+};
+
+export default ContactPage;
