+++ pages/about.js
@@ -1,4 +1,5 @@
 import Footer from "@components/footer";
+import AboutPage from "@components/pageSections/aboutPage";
 import BannerAbout from "@components/pageSections/aboutPage/bannerSection";
 import HistorySection from "@components/pageSections/aboutPage/historySection";
 import OfficesSection from "@components/pageSections/aboutPage/officesSection";
@@ -8,19 +9,13 @@ import TeachersSection from "@components/pageSections/aboutPage/teachersSection"
 import WorkValues from "@components/pageSections/aboutPage/workValues";
 import React from "react";
 
-const AboutPage = () => {
+const About = () => {
   return (
     <>
-      <BannerAbout />
-      <Achievements />
-      <StorySection />
-      <WorkValues />
-      <TeachersSection />
-      <HistorySection />
-      <OfficesSection />
+      <AboutPage/>
       <Footer />
     </>
   );
 };
 
-export default AboutPage;
+export default About;
