+++ pages/blog.js
@@ -1,17 +1,10 @@
-import Footer from "@components/footer";
-import BlogBanner from "@components/pageSections/blogPage/blogBanner";
-import BlogCoursesSection from "@components/pageSections/blogPage/blogCourses";
-import BlogSection from "@components/pageSections/blogPage/blogSection";
-import SubscribeToday from "@components/pageSections/blogPage/subscribeToday";
+import BlogPage from "@components/pageSections/blogPage";
 import React from "react";
 
 const Blog = () => {
   return (
     <>
-      <BlogBanner />
-      <BlogSection />
-      <SubscribeToday />
-      <BlogCoursesSection />
+      <BlogPage/>      
     </>
   );
 };
