+++ pages/contact.js
@@ -1,13 +1,10 @@
-import AskedQues from "@components/pageSections/contactPage/askedQues";
-import AskQuesList from "@components/pageSections/contactPage/askedQues/askedQuesData";
-import ContactForm from "@components/pageSections/contactPage/contactForm";
+import ContactPage from "@components/pageSections/contactPage";
 import React from "react";
 
 const Contact = () => {
   return (
     <>
-      <ContactForm />
-      <AskQuesList />
+      <ContactPage/>
     </>
   );
 };
