+++ pages/index.js
@@ -1,8 +1,11 @@
 import Layout from "@components/layouts";
+import HomePage from "@components/pageSections/homePage";
 import React from "react";
 
 const Home = () => {
-  return <></>;
+  return <>
+  <HomePage/>
+  </>;
 };
 Home.getLayout = function getLayout(page) {
   return <Layout>{page}</Layout>;
