+++ pages/individualcourse.js
@@ -1,12 +1,10 @@
 import React from "react";
 import IndividualCourse from "@components/pageSections/individualCourse";
-import Footer from "@components/footer";
 
 const Individual = () => {
   return (
     <>
       <IndividualCourse />
-      <Footer />
     </>
   );
 };
